staging: vchiq_arm: Get the rid off struct vchiq_2835_state
The whole benefit of this encapsulating struct is questionable.
It just stores a flag to signalize the init state of vchiq_arm_state.
Beside the fact this flag is set too soon, the access to uninitialized
members should be avoided per design. So initialize vchiq_arm_state
properly before assign it directly to vchiq_state.

Signed-off-by: Stefan Wahren <wahrenst@gmx.net>