pmdomain: raspberrypi-power: Never turn off V3D power domain
The V3D power domain is essential to the vc4 DRM driver. As soon
this power domain is turned on by the driver, this should never be
turned off. Otherwise this triggers the GPU hang detection, which
results in an unwanted reset:

[  165.791899] [drm] Resetting GPU.

Since the V3D power domain isn't enabled during init phase, it's
not possible to use GENPD_FLAG_ALWAYS_ON here.

Signed-off-by: Stefan Wahren <wahrenst@gmx.net>